Rename the actual UI pie menu files
We changed the name of the class, so let's update the filenames to
match.

Re: issue #123
